sp-consumer-dish-service 套餐分页查询、起售停售套餐。修改实体类Setmeal使用bulid注解和无参、全参构造器注解。

+++ sp-common/sp-common-core/src/main/java/com/sp/core/constants/SystemConstants.java
@@ -14,11 +14,11 @@ public final class SystemConstants {
     /**
      * 分页大小
      */
-    public static final int PAGE_SIZE = 8;
+    public static final Long PAGE_SIZE = 8L;
     /**
      * 默认当前页
      */
-    public static final int PAGE_NO = 1;
+    public static final Long PAGE_NO = 1L;
     //项目名
     public static final String PROJECT_NAME = "SummerPockets";
